+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-* Implement the `view` feature of the get method.
+* Implement the `view` feature of the get and iterate method.
+* Implement the `filter_by_formula` feature of the iterate method.
 * Raises an error when trying to access a table that was not created before.
 
 ## v0.0.4 [2018-06-19]